name and directory reorganisation
