+++ src/helpers/string-formatter.js
