+++ src/schemas/library-schema.json
