+++ test/validator.test.js
@@ -1,7 +1,7 @@
 const fs = require('fs-extra');
 const path = require('path');
 
-const H5pPackageValidator = require('../src/validator');
+const H5pPackageValidator = require('../src/package-validator');
 const H5PConfig = require('../src/config');
 const TranslationService = require('../src/translation-service');
